consolidate burn action into a single method
